requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Machine Learning
torch==2.1.1
scikit-learn==1.3.2

# Quantum Computing
qiskit==0.45.1
qiskit-aer==0.13.1
qiskit-ibm-runtime==0.15.0

# Visualization
matplotlib==3.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
hypothesis==6.92.1

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging
python-json-logger==2.0.7
